+++ core/src/main/kotlin/fragments/NodeFragment.kt
@@ -1,6 +1,8 @@
 package dev.inmo.navigation.core.fragments
 
+import android.os.Bundle
 import androidx.fragment.app.Fragment
+import dev.inmo.navigation.core.NavigationNodeId
 import dev.inmo.navigation.core.configs.NavigationNodeDefaultConfig
 import kotlinx.coroutines.flow.MutableStateFlow
 
@@ -11,6 +13,27 @@ abstract class NodeFragment<Config : Base, Base : NavigationNodeDefaultConfig> :
     protected var config: Config
         get() = _configState.value
         set(value) { _configState.value = value }
+    internal var configId: String
+        get() = runCatching { config.id }.getOrElse { arguments ?.getString("configId") ?: throw it }
+        set(value) {
+            (arguments ?: Bundle().apply { arguments = this }).putString("configId", value)
+        }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        savedInstanceState ?.let {
+            arguments = Bundle().apply {
+                putAll(it)
+                putAll(arguments ?: return@apply)
+            }
+        }
+        configId = arguments ?.getString("configId") ?: config.id
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onSaveInstanceState(outState: Bundle) {
+        super.onSaveInstanceState(outState)
+        outState.putString("configId", configId)
+    }
 
     internal fun setNode(
         node: AndroidFragmentNode<Config, Base>,
@@ -18,5 +41,6 @@ abstract class NodeFragment<Config : Base, Base : NavigationNodeDefaultConfig> :
     ) {
         this.node = node
         this._configState = configStateFlow
+        configId = configStateFlow.value.id
     }
 }